==> notebooks/1-expression-preprocessing.ipynb <==
import pandas as pd
import numpy as np
from sklearn.preprocessing import FunctionTransformer, MinMaxScaler

Create dataframe by transposing matrix

expression_df = pd.read_csv('../data/raw/POG570_TPM_matrix.txt', index_col=0, sep='\t')

expression_df = expression_df.T

expression_df.to_csv('../data/raw/POG570_TPM.tsv', sep='\t')

## Raw data

expression_raw_df = pd.read_csv('../data/raw/POG570_TPM.tsv', sep='\t', index_col = 0)

## Scale to range [0, 1]

minmax_scaler = MinMaxScaler()
expression_minmax_matrix = minmax_scaler.fit_transform(expression_raw_df.iloc[:, 1:])

expression_minmax_df = pd.DataFrame(expression_minmax_matrix, columns=expression_raw_df.iloc[:, 1:].columns.get_values())

expression_minmax_df.insert(0, 'pog_id', expression_raw_df.index.get_values())

expression_minmax_df.to_csv('../data/processed/expression_tpm_minmax.tsv', sep='\t')

## Transform to Logarithmic scale
According to https://www.biorxiv.org/content/early/2018/07/11/364323

def transform_function(x):
    return np.log1p(x) if np.log1p(x) >=1 else 0
transform_funct_vect = np.vectorize(transform_function)

transformer = FunctionTransformer(transform_funct_vect, validate=True)

expression_log_df = pd.DataFrame(transformer.transform(expression_raw_df.iloc[:, 1:]), columns=expression_raw_df.iloc[:, 1:].columns.get_values())#, index=expr_data.index)

expression_log_df.insert(0, 'pog_id', expression_raw_df.index.get_values())

expression_log_df.to_csv('../data/processed/expression_tpm_log.tsv', sep='\t')

transformer.transform(expression_raw_df.iloc[:, 1:])

## Apply [0, 1] scale after log 

expression_log_minmax_matrix = minmax_scaler.fit_transform(expression_log_df.iloc[:, 1:])

expression_log_minmax_df = pd.DataFrame(expression_log_minmax_matrix, columns=expression_raw_df.iloc[:, 1:].columns.get_values())

expression_log_minmax_df.insert(0, 'pog_id', expression_raw_df.index.get_values())

expression_log_minmax_df.to_csv('../data/processed/expression_tpm_log_minmax.tsv', sep='\t')

expression_log_minmax_df.to_hdf('../data/hdf5/expresion_tpm_log_minmax.tsv', key='log_minmax')
