==> notebooks/2-expression-feature-selection.ipynb <==
import pandas as pd
import numpy as np

Select 299 driver genes according to https://www.ncbi.nlm.nih.gov/pubmed/29625053

table = pd.read_csv('../data/external/table_s1.csv', header=2)

cancer_genes = np.unique(table['Gene'])

expression_log_df = pd.read_csv('../data/processed/expression_tpm_log.tsv', sep='\t', index_col=0)
expression_minmax_df = pd.read_csv('../data/processed/expression_tpm_minmax.tsv', sep='\t', index_col=0)

mapping_df = pd.read_csv('../data/external/gene_mapping_and_hugo_ids.txt', sep='\t', index_col=0)

columns_to_drop = []
missing_vals = 0
for ensg_gene in expression_log_df.iloc[:, 1:].columns:
    if ensg_gene not in mapping_df.index.values:
        columns_to_drop.append(ensg_gene)
        missing_vals += 1
        continue
    if mapping_df.loc[ensg_gene, 'HUGO'] not in cancer_genes:
        columns_to_drop.append(ensg_gene)
print(missing_vals)

len(columns_to_drop)

len(expression_log_df.columns)

expression_log_df.columns

expression_log_df = expression_log_df.drop(columns_to_drop, axis=1)

expression_minmax_df = expression_minmax_df.drop(columns_to_drop, axis=1)

expression_log_df.to_csv('../data/processed/expression_tpm_log_selected.tsv', sep='\t')

expression_minmax_df.to_csv('../data/processed/expression_tpm_minmax_selected.tsv', sep='\t')
