==> notebooks/3-svm.ipynb <==
import pandas as pd
import numpy as np

from sklearn.svm import SVC

random_seed = 42

## Get data

drugs_df = pd.read_csv('../data/processed/drugs_filtered.tsv', sep='\t', index_col=0)

drugs_df = drugs_df[['pog_id', 'drug_name', 'cancer_cohort', 'zscore', 'days_on_tx_since_biopsy']]

# expression_df = pd.read_csv('../data/processed/expression_tpm_log_selected.tsv', sep='\t', index_col=0)
expression_df = pd.read_csv('../data/processed/expression_tpm_minmax_selected.tsv', sep='\t', index_col=0)

## Discretize drug z-scores by K bins

from sklearn.preprocessing import KBinsDiscretizer

discretizer = KBinsDiscretizer(encode='ordinal', n_bins=2, strategy='kmeans')

binned_zscores = discretizer.fit_transform(drugs_df['days_on_tx_since_biopsy'].values.reshape(-1, 1))[:, 0]

drugs_df['binned_days'] = binned_zscores

# 1. Naive approach: all cancer types, all drugs

## Get dummy variables for drug names

drug_dummies = pd.get_dummies(drugs_df['drug_name'])

## Join drug data, drug dummies and expression data

expression_df = expression_df.set_index('pog_id')

drugs_df = drugs_df.join(drug_dummies)

drugs_expression_df = drugs_df.join(expression_df, on='pog_id', how='inner', sort=True)

## Prepare features and labels for SVM

X_columns = np.append(expression_df.columns.values, drug_dummies.columns.values)

X = drugs_expression_df.loc[:, X_columns]

y = drugs_expression_df['binned_zscores']

### Train test split

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.33, random_state=random_seed)

## SVM

naive_svm = SVC(C=1, kernel='linear')

naive_svm.fit(X_train, y_train)

y_pred = naive_svm.predict(X_test)

## Binary classification metrics

from sklearn.metrics import confusion_matrix, f1_score

confusion_matrix(y_test, y_pred)

tn, fp, fn, tp = confusion_matrix(y_test, y_pred).ravel()

tn, fp, fn, tp

f1_score(y_pred=y_pred, y_true=y_test)

# 2. All cancer types, only GEMCITABINE or FLUOROURACIL

drugs_filtered_df = drugs_df[drugs_df['drug_name'] == 'CISPLATIN']

drugs_expression_filtered_df = drugs_filtered_df.join(expression_df, on='pog_id', how='inner', sort=True)

drugs_expression_filtered_df = drugs_expression_filtered_df[drugs_expression_filtered_df['cancer_cohort'] == 'LUNG']

X = drugs_expression_filtered_df.loc[:, expression_df.columns.values]

y = drugs_expression_filtered_df['binned_zscores']

drugs_expression_filtered_df

## Train test split

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=random_seed)

## SVM

filtered_svm = SVC(C=0.01, kernel='linear')

filtered_svm.fit(X_train, y_train)

y_pred = filtered_svm.predict(X_test)

## Binary classification metrics

from sklearn.metrics import confusion_matrix, f1_score

confusion_matrix(y_test, y_pred)

tn, fp, fn, tp = confusion_matrix(y_test, y_pred).ravel()

tn, fp, fn, tp

f1_score(y_pred=y_pred, y_true=y_test)
