==> notebooks/5-svm-rfe.ipynb <==
_random_seed_ = 42

import pandas as pd
import numpy as np
import math
from matplotlib import pyplot as plt

expression_minmax_df = pd.read_csv('../data/processed/expression_log_tpm_minmax.tsv', sep='\t', index_col=0)

expression_minmax_df = expression_minmax_df.set_index('pog_id')

drugs_df = pd.read_csv('../data/processed/drugs_filtered.tsv', sep='\t', index_col=0)

drugs_selected_df = drugs_df[['pog_id', 'drug_name', 'days_on_tx_since_biopsy', 'cancer_cohort']]

# Prepare features and labels

## Join drugs and expression tables

drugs_expression_df = drugs_selected_df.join(expression_minmax_df, on='pog_id', how='inner')

drugs_expression_df = drugs_expression_df.drop_duplicates()

Number of drug types and their names

## Select cancer type and drug
decide based on notebook 0

cancer_types = np.unique(drugs_expression_df['cancer_cohort'])

drug_names = np.unique(drugs_expression_df['drug_name'])

# drugs_expression_sel_df = drugs_expression_df[(drugs_expression_df['cancer_cohort'] == 'BRCA') & (drugs_expression_df['drug_name'] == 'CAPECITABINE')]
drugs_expression_sel_df = drugs_expression_df[(drugs_expression_df['cancer_cohort'] == 'LYMP')]

# testing

math.floor(0.25*len(drugs_expression_sel_df['pog_id']))

len(drugs_expression_sel_df['pog_id'])

select_index = np.random.choice(np.unique(drugs_expression_sel_df['pog_id']), size=4)

mask = drugs_expression_sel_df['pog_id'].isin(select_index)

drugs_expression_sel_df[mask]

len(drugs_expression_sel_df['pog_id'])

test = drugs_expression_sel_df.drop_duplicates()

## Set features (X) and labels (y)

X = drugs_expression_sel_df.loc[:, expression_minmax_df.columns]

y = drugs_expression_sel_df.loc[:, 'days_on_tx_since_biopsy']

expression_minmax_df

## Power transform on y

from sklearn.preprocessing import PowerTransformer

power_transformer = PowerTransformer(method='box-cox', standardize=True)

y_trans = power_transformer.fit_transform(y.values.reshape(-1, 1))[:, 0]

## Naive feature selection: Variance threshold

from sklearn.feature_selection import VarianceThreshold

variance_selector = VarianceThreshold(threshold=0.01)

X_selected = variance_selector.fit_transform(X)

X_selected.shape

selected_columns = X.columns[variance_selector.get_support()]

X_selected_df = pd.DataFrame(data=X_selected, columns=selected_columns, index=X.index)

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X_selected_df, y_trans, test_size=0.25, random_state=_random_seed_)
# X_train, X_test, y_train, y_test = train_test_split(X, y_trans, test_size=0.25, random_state=_random_seed_)

# Recursive Feature Selection

from sklearn.feature_selection import RFE

from sklearn.svm import LinearSVR

num_features = math.floor(len(X_selected_df.columns)/20)

num_features = 2000

svr = LinearSVR(max_iter=10000)

selector = RFE(svr, num_features, step=0.01, verbose=0)

selector = selector.fit(X_train, y_train)

selector.score(X_test, y_test)

svr.fit(X_train, y_train)

svr.score(X_test, y_test)

drugs_expression_df['drug_name'] == 'FLUOROURACIL'

test = [(1,2), (3,4), (5,6)]
test2 = [1,2,3]

for i, j in zip(test, test2):
    a, b = i
    print(a,b)
